Reverted changes made in r311310.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@50365 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objmgr/seq_map_ci.cpp
@@ -424,7 +424,7 @@ bool CSeqMap_CI::x_Push(TSeqPos pos, bool resolveExternal)
         if ( (m_Stack.size() & 63) == 0 ) {
             // check for self-recursion every 64'th stack frame
             const CSeqMap* top_seq_map = &m_Stack.back().x_GetSeqMap();
-            for ( size_t i = m_Stack.size()-2; i >= 0; --i ) {
+            for ( int i = m_Stack.size()-2; i >= 0; --i ) {
                 if ( &m_Stack[i].x_GetSeqMap() == top_seq_map ) {
                     NCBI_THROW(CSeqMapException, eSelfReference,
                                "Self-reference in CSeqMap");
